backend/requirements.txt
torch>=2.2.0
transformers>=4.44.0
peft>=0.12.0
accelerate>=0.34.0
Pillow>=10.0.0
bitsandbytes>=0.43.0
tensorboard>=2.17.0
python-dotenv>=1.0.0
